Optimize complexity and files length. Cleaned code.
